move existing states to Phone specific namespace
