+++ DuckDuckGo/OmniBarState.swift
@@ -39,123 +39,3 @@ protocol OmniBarState {
     var onBrowsingStartedState: OmniBarState { get }
     var onBrowsingStoppedState: OmniBarState { get }
 }
-
-struct HomeEmptyEditingState: OmniBarState {
-    var clearTextOnStart = true
-    var allowsTrackersAnimation = false
-    var showSearchLoupe = true
-    let showSiteRating = false
-    let showBackground = false
-    let showClear = false
-    let showRefresh = false
-    let showMenu = false
-    let showSettings = false
-    let showCancel: Bool = true
-    var name: String { return Type.name(self) }
-    var onEditingStoppedState: OmniBarState { return HomeNonEditingState() }
-    var onEditingStartedState: OmniBarState { return self }
-    var onTextClearedState: OmniBarState { return self }
-    var onTextEnteredState: OmniBarState { return HomeTextEditingState() }
-    var onBrowsingStartedState: OmniBarState { return BrowsingNonEditingState() }
-    var onBrowsingStoppedState: OmniBarState { return self }
-}
-
-struct HomeTextEditingState: OmniBarState {
-    var clearTextOnStart = false
-    var allowsTrackersAnimation = false
-    var showSearchLoupe = true
-    let showSiteRating = false
-    let showBackground = false
-    let showClear = true
-    let showRefresh = false
-    let showMenu = false
-    let showSettings = false
-    let showCancel: Bool = true
-    var name: String { return Type.name(self) }
-    var onEditingStoppedState: OmniBarState { return HomeNonEditingState() }
-    var onEditingStartedState: OmniBarState { return self }
-    var onTextClearedState: OmniBarState { return HomeEmptyEditingState() }
-    var onTextEnteredState: OmniBarState { return self }
-    var onBrowsingStartedState: OmniBarState { return BrowsingNonEditingState() }
-    var onBrowsingStoppedState: OmniBarState { return HomeEmptyEditingState() }
-}
-
-struct HomeNonEditingState: OmniBarState {
-    var clearTextOnStart = true
-    var allowsTrackersAnimation = false
-    var showSearchLoupe = true
-    let showSiteRating = false
-    let showBackground = true
-    let showClear = false
-    let showRefresh = false
-    let showMenu = false
-    let showSettings = true
-    let showCancel: Bool = false
-    var name: String { return Type.name(self) }
-    var onEditingStoppedState: OmniBarState { return self }
-    var onEditingStartedState: OmniBarState { return HomeEmptyEditingState() }
-    var onTextClearedState: OmniBarState { return HomeEmptyEditingState() }
-    var onTextEnteredState: OmniBarState { return HomeTextEditingState() }
-    var onBrowsingStartedState: OmniBarState { return BrowsingNonEditingState() }
-    var onBrowsingStoppedState: OmniBarState { return HomeNonEditingState() }
-}
-
-struct BrowsingEmptyEditingState: OmniBarState {
-    var clearTextOnStart = true
-    var allowsTrackersAnimation = false
-    var showSearchLoupe = true
-    let showSiteRating = false
-    let showBackground = false
-    let showClear = false
-    let showRefresh = false
-    let showMenu = false
-    let showSettings = false
-    let showCancel: Bool = true
-    var name: String { return Type.name(self) }
-    var onEditingStoppedState: OmniBarState { return BrowsingNonEditingState() }
-    var onEditingStartedState: OmniBarState { return self }
-    var onTextClearedState: OmniBarState { return self }
-    var onTextEnteredState: OmniBarState { return BrowsingTextEditingState() }
-    var onBrowsingStartedState: OmniBarState { return self }
-    var onBrowsingStoppedState: OmniBarState { return HomeEmptyEditingState() }
-}
-
-struct BrowsingTextEditingState: OmniBarState {
-    var clearTextOnStart = false
-    var allowsTrackersAnimation = false
-    var showSearchLoupe = true
-    let showSiteRating = false
-    let showBackground = false
-    let showClear = true
-    let showRefresh = false
-    let showMenu = false
-    let showSettings = false
-    let showCancel: Bool = true
-    var name: String { return Type.name(self) }
-    var onEditingStoppedState: OmniBarState { return BrowsingNonEditingState() }
-    var onEditingStartedState: OmniBarState { return self }
-    var onTextClearedState: OmniBarState { return BrowsingEmptyEditingState() }
-    var onTextEnteredState: OmniBarState { return self }
-    var onBrowsingStartedState: OmniBarState { return self }
-    var onBrowsingStoppedState: OmniBarState { return HomeEmptyEditingState() }
-}
-
-struct BrowsingNonEditingState: OmniBarState {
-    var clearTextOnStart = false
-    var allowsTrackersAnimation = true
-    var showSearchLoupe = false
-    let showSiteRating = true
-    let showBackground = true
-    let showClear = false
-    let showRefresh = true
-    let showMenu = true
-    let showSettings = false
-    let showCancel: Bool = false
-    var name: String { return Type.name(self) }
-    var onEditingStoppedState: OmniBarState { return self }
-    var onEditingStartedState: OmniBarState { return BrowsingTextEditingState() }
-    var onTextClearedState: OmniBarState { return BrowsingEmptyEditingState() }
-    var onTextEnteredState: OmniBarState { return BrowsingTextEditingState() }
-    var onBrowsingStartedState: OmniBarState { return self }
-    var onBrowsingStoppedState: OmniBarState { return HomeNonEditingState() }
-}
